trigger login modal if user attempts to upload a bank statement before they are logged in
